+++ packages/gamut-styles/src/variables/timing.ts
@@ -1,11 +1,13 @@
 export const timingValues = {
   fast: 150,
   medium: 200,
+  base: 300,
   slow: 350,
 };
 
 export const timing = {
   fast: `${timingValues.fast}ms`,
   medium: `${timingValues.medium}ms`,
+  base: `${timingValues.base}ms`,
   slow: `${timingValues.slow}ms`,
 };