+++ packages/gamut/src/Tip/__tests__/ToolTip.test.tsx
@@ -1,4 +1,5 @@
 import { setupRtl } from '@codecademy/gamut-tests';
+import { waitFor } from '@testing-library/dom';
 import userEvent from '@testing-library/user-event';
 
 import { ToolTipMock } from './mocks';
@@ -83,7 +84,7 @@ describe('floating placement', () => {
     view.getByRole('button', { name: 'Click' });
     expect(view.getByRole('tooltip', { hidden: true })).toHaveTextContent(info);
   });
-  it('shows the tip when it is hovered over', () => {
+  it('shows the tip when it is hovered over', async () => {
     const { view } = renderView({
       placement: 'floating',
     });
@@ -93,7 +94,7 @@ describe('floating placement', () => {
     userEvent.hover(view.getByRole('button'));
 
     view.getByRole('tooltip', { hidden: true });
-    expect(view.queryAllByText(info).length).toBe(2);
+    await waitFor(() => expect(view.queryAllByText(info).length).toBe(2));
   });
   it('calls onClick when clicked', () => {
     const { view } = renderView({});